Bump the nng submodule to nng v1.11 (#4)
The changes include:
  - Updating the nng submodule to v1.11
  - Fixing header includes to use angle brackets instead of quotes (making them use compiler search path)
  - Updating bindgen from 0.54 to 0.72 and adjust it's usage
  - Adding TLS library linking (mbedcrypto, mbedtls, mbedx509)
  - Building only the `nng` target instead of the whole workspace
  - Adding a runtime version verification test

The main motivation is keeping up with upstream nng development and
getting new features/fixes from v1.11. The build system changes (header
paths, cmake target selection) are necessary to work with the new nng
version and to make the build more flexible (allowing use of system nng
instead of just the submodule). The bindgen is required for
compatibility with newer Rust toolchains.

This also add a basic CI configuration for GitHub actions. The
pipelines are executed on the GH linux runner only and are a subject to
be extended later with more platforms and more sophisticated checks. Also
it currently lacks check of `rustfmt`, `cargo hack`, `cargo deny` etc.

Co-authored-by: Felix Obenhuber <felix.obenhuber@helsing.ai>