refactor(nng-sys): remove cmake generator features (#14)
Remove the cmake-* generator features. It's not generally the job of a
`-sys` crate to dictate the means-of-build, that should be left to the
user further "up". For example, the CMake generator is generally
controlled globally through the `CMAKE_GENERATOR` environment variable,
which the `cmake` crate already respects.

This also simplifies the crate feature list and eliminates maintenance
burden for tracking new Visual Studio versions. Not to mention it makes
CI simpler since there are fewer "weird" combinations to try/skip.