nng-sys: revamp features + discover system NNG (#16)
Previously, without `build-nng`, we just emitted
`cargo:rustc-link-lib=dylib=nng` and hoped for the best. This wasn't
great for anyone who already had NNG built elsewhere, such as from their
distro, from nix, or even from a local dev build of NNG that they might
be testing.

The new discovery logic implements a preference hierarchy of where to
pick up NNG from (when not using a vendored build, which remains the
default):

1. Explicit env vars (`NNG_DIR`, `NNG_LIB_DIR`+`NNG_INCLUDE_DIR`)
2. pkg-config (though nanomsg/nng#926, so this is to future-proof) / vcpkg
3. Platform-specific fallback paths ("-lnng and pray" or Homebrew on macOS)

Vendoring is still enabled through a feature and on by default. This is
mainly because NNG isn't _generally_ going to be installed on users
systems, unlike more widely used libraries like openssl or zlib.
Besides, NNG is fast and easy enough to build that it's not too much of
a problem in practice. However, to allow end users to force the use of a
system-wide NNG install, I've added support for the `NNG_NO_VENDOR`
environment variable to override that behavior. This matches the pattern
from openssl-sys (`OPENSSL_NO_VENDOR`) and libgit2-sys
(`LIBGIT2_NO_VENDOR`), and is useful for deployments where deep control
over exact versions of dependencies are needed (e.g., in systems that
need to deal with compliance, SBOMs, etc.).

I've introduced another environment variable, `NNG_STATIC`, which can be
set to force either static or dynamic linking of NNG, but that one is
less interesting. There's also a `static` feature to request static
linking specifically (which would only be relevant when not using
vendored, where static is the default).

To be consistent with the rest of the Rust -sys ecosystem, I've renamed
the `build-nng` feature to just `vendored` (like libsqlite3-sys and
openssl-sys). Similarly, I flipped no_std to be an opt-in (and on by
default) `std` feature; this is needed to make the feature additive,
plus
<https://www.reddit.com/r/rust/comments/1hs6spy/psa_for_std_feature_in_no_std_libraries/>.

I've also renamed some of the other features:

- build-bindgen → bindgen, since we're not "building" bindgen
- nng-stats → vendored-stats, since it only affects vendored CMake build
- nng-* → * without the prefix, since it's implied that this is for nng

One important and more subtle change is that we now always re-generate
the bindings when using the system-provided NNG, since there is no
guarantee that it is the same version (or built with the same features)
as the pre-generated bindings. Vendored builds continue using
pre-generated bindings (unless the `bindgen` feature is enabled
explicitly or implicitly). In addition, I've added a CI job that checks
that the checked-in bindings are up to date.